+++ test/typescript/querycursor.ts
@@ -3,7 +3,6 @@ import { Schema, model, Document, Types } from 'mongoose';
 const schema: Schema = new Schema({ name: { type: 'String' } });
 
 interface ITest extends Document {
-  _id?: Types.ObjectId,
   name?: string;
 }